Hotfix for potential script exceptions
The file name the host got back when there were compilation errors was
for the temp file created by the precompiler, but since the temp files
are cleared after compilation, there was no reference to the script that
caused the error at all. This replaces the name of the temp file with
the name of the original script.